requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
python-dateutil>=2.8.2
pytz>=2023.3
pydantic>=2.0.0
